Safari-609.1.11.1/LayoutTests/js/dom/modules/module-execution-order-mixed.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test module execution order between mixed tags.');
debug('We keep the execution order between modules.');
window.count = 0;

// Module will be executed asynchronously.
window.jsTestIsAsync = true;
debug('Module is not executed yet.');
</script>
<script src="../../../resources/js-test-post.js"></script>
<script type="module" src="resources/module-execution-order-mixed.js"></script>
<script type="module">
shouldBe("count++", "4");
</script>
<script type="module" src="resources/module-execution-order-mixed-2.js"></script>
<script type="module">
shouldBe("count++", "6");
finishJSTest();
</script>
</body>
</html>
